+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* Adding function to increment the badge number
+* Adding function to decrement the badge number
+* Updating dependencies
+* Updating example app
+* Updating native code to respect latest versions of Android, iOS and macOS
+
 ## 1.0.1
 
 * Bug fix with android